requirements.txt
requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
python-dotenv
tiktoken
watchdog